+++ integration-libs/cdc/core/auth/facade/index.ts
@@ -1,0 +1,1 @@
+export * from './cdc-auth.service';
