+++ integration-libs/cdc/core/auth/store/effects/cdc-user-token.effect.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
 import { Actions, Effect, ofType } from '@ngrx/effects';
-import { CdcAuthActions, CdcAuthService } from '@spartacus/cdc/root';
 import {
   GlobalMessageService,
   GlobalMessageType,
@@ -8,7 +7,9 @@ import {
 } from '@spartacus/core';
 import { EMPTY, Observable, of } from 'rxjs';
 import { catchError, map, mergeMap, switchMap } from 'rxjs/operators';
+import { CdcAuthService } from '../../facade';
 import { CdcUserAuthenticationTokenService } from '../../services/user-authentication/cdc-user-authentication-token.service';
+import { CdcAuthActions } from '../actions';
 
 @Injectable()
 export class CdcUserTokenEffects {
