+++ integration-libs/cdc/core/index.ts
@@ -1,2 +1,4 @@
+export * from './auth/facade/index';
+export * from './auth/store/actions/index';
 export * from './cdc-core.module';
 export * from './models/cms.model';
