+++ integration-libs/cdc/core/ng-package.json
@@ -4,10 +4,7 @@
     "entryFile": "./public_api.ts",
     "umdModuleIds": {
       "@spartacus/core": "core",
-      "@spartacus/storefront": "storefront",
-      "@spartacus/asm/root": "asm/root",
-      "@ngrx/store": "store",
-      "@ngrx/effects": "effects"
+      "@spartacus/storefront": "storefront"
     }
   }
 }
