+++ integration-libs/cdc/ng-package.json
@@ -6,7 +6,10 @@
     "umdModuleIds": {
       "@spartacus/core": "core",
       "@spartacus/storefront": "storefront",
-      "rxjs": "rxjs"
+      "@spartacus/asm/root": "asm/root",
+      "rxjs": "rxjs",
+      "@ngrx/store": "store",
+      "@ngrx/effects": "effects"
     }
   }
 }
