+++ integration-libs/cdc/root/facade/index.ts
@@ -1,2 +1,1 @@
-export * from './cdc-auth.service';
 export * from './cdc-js.service';
