+++ integration-libs/cdc/root/ng-package.json
@@ -4,9 +4,7 @@
     "entryFile": "./public_api.ts",
     "umdModuleIds": {
       "@spartacus/core": "core",
-      "@spartacus/storefront": "storefront",
-      "@spartacus/asm/root": "asm/root",
-      "@ngrx/store": "store"
+      "@spartacus/storefront": "storefront"
     }
   }
 }
